Internal #861: Refactor Window Naive Aggregator
Move the window Naive Aggregator class to separate files.
