+++ src/function/window/window_aggregate_function.cpp
@@ -1,4 +1,5 @@
 #include "duckdb/function/window/window_aggregate_function.hpp"
+#include "duckdb/function/window/window_naive_aggregator.hpp"
 #include "duckdb/function/window/window_shared_expressions.hpp"
 #include "duckdb/planner/expression/bound_reference_expression.hpp"
