+++ src/include/duckdb/function/window/window_segment_tree.hpp
@@ -19,18 +19,6 @@
 
 namespace duckdb {
 
-// Used for validation
-class WindowNaiveAggregator : public WindowAggregator {
-public:
-	WindowNaiveAggregator(const BoundWindowExpression &wexpr, const WindowExcludeMode exclude_mode,
-	                      WindowSharedExpressions &shared);
-	~WindowNaiveAggregator() override;
-
-	unique_ptr<WindowAggregatorState> GetLocalState(const WindowAggregatorState &gstate) const override;
-	void Evaluate(const WindowAggregatorState &gsink, WindowAggregatorState &lstate, const DataChunk &bounds,
-	              Vector &result, idx_t count, idx_t row_idx) const override;
-};
-
 class WindowConstantAggregator : public WindowAggregator {
 public:
 	WindowConstantAggregator(const BoundWindowExpression &wexpr, WindowExcludeMode exclude_mode_p,
